requirements.txt
# Core dependencies
poke-env>=0.4.0
numpy>=1.24.0
torch>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0

